Add Mongo DB & log volumes
This adds the database & logs to volumes.  The log no longer shows
in the system output.  Binding would be more useful to see the log
on the host environment.  However, all the ways that I've tried, I
have gotten permission errors when binding.  Eventually, I found
out that this is due to the file systems not being capable.