Update node server to use mongodb client instead
The problem with Mongoose was I would have to define the schemas
twice: once in the database init scripts and once for Mongoose.
I did not want to do that.  I prefer only defining them in the
database.  Mongodb client allows me to do that.

Also, added testing for retrieving an image from the database.