Add React frontend & rename node folder to backend
First, I renamed the node folder to backend & added a folder
for the frontend.  I followed the steps to initialize a React
application.  I had a lot of trouble with the builds timing out
so I added lock-files for both the frontend & the backend.  Then
for my application to work, I needed routers & I had a lot of
trouble with that.  The end result is the start to my frontend
part of my application.  I have stubs to three pages: welcome,
create game, and play game.