Add backend logic to create game
This creates the schema for a game in the database.  In the
backend server, I added a path to create a game.  The path
accepts types of "picture", "word", or "both" and sizes of
"5x4" or "5x5."  Depending on the type and size, it will
create an array of object ids to the assets & a color
grid to match.  Number of colors are pulled from the
official codenames rules.  The result is a game document
will be added to the database and if successful, a successful
JSON will show up.  On an error, a non-successful JSON will
show up.