Fix css MIME type in production
Production's css files failed to load because the MIME type was
text/plain, not text/css.  This was because I changed too much of
the nginx.conf file.  I only needed to change the port which is
in the server section.  However, I changed the full file before
and removed that line.  The full file includes all the files in
the configuration directory.  Therefore, the new fix only adds
a file in the directory with only the server part changed.  This
is different from the rev_proxy version where limited configuration
is appropriate.